+++ protocol/synthetix/contracts/interfaces/IRewardsManagerModule.sol
@@ -5,23 +5,32 @@ pragma solidity ^0.8.0;
  * @title Module for connecting rewards distributors to vaults.
  */
 interface IRewardsManagerModule {
+    /**
+     * @notice Emitted when the pool owner or an existed distributor sets up rewards for vault participants.
+     */
     event RewardsDistributed(
         uint128 indexed poolId,
         address indexed collateralType,
         address distributor,
-        uint amount,
-        uint start,
-        uint duration
+        uint256 amount,
+        uint256 start,
+        uint256 duration
     );
 
+    /**
+     * @notice Emitted when a vault participant claims rewards.
+     */
     event RewardsClaimed(
         uint128 indexed accountId,
         uint128 indexed poolId,
         address indexed collateralType,
         address distributor,
-        uint amount
+        uint256 amount
     );
 
+    /**
+     * @notice Emitted when a new rewards distributor is registered.
+     */
     event RewardsDistributorRegistered(
         uint128 indexed poolId,
         address indexed collateralType,
@@ -43,7 +52,7 @@ interface IRewardsManagerModule {
     function distributeRewards(
         uint128 poolId,
         address collateralType,
-        uint amount,
+        uint256 amount,
         uint64 start,
         uint32 duration
     ) external;
@@ -56,7 +65,7 @@ interface IRewardsManagerModule {
         address collateralType,
         uint128 accountId,
         address distributor
-    ) external returns (uint);
+    ) external returns (uint256);
 
     /**
      * @notice For a given position, return the rewards that can currently be claimed
@@ -65,7 +74,7 @@ interface IRewardsManagerModule {
         uint128 poolId,
         address collateralType,
         uint128 accountId
-    ) external returns (uint[] memory, address[] memory);
+    ) external returns (uint256[] memory, address[] memory);
 
     /**
      * @notice Returns the number of individual units of amount emitted per second per share for the given poolId, collateralType, distributor vault.
@@ -74,5 +83,5 @@ interface IRewardsManagerModule {
         uint128 poolId,
         address collateralType,
         address distributor
-    ) external view returns (uint);
+    ) external view returns (uint256);
 }
